pyproject.toml
[project]
name = "langgraph-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "arize-phoenix>=11.32.1",
    "langchain-ollama>=0.3.8",
    "langchain[anthropic]>=0.3.27",
    "langgraph>=0.6.7",
    "openinference-instrumentation-langchain>=0.1.52",
    "opentelemetry-exporter-otlp>=1.37.0",
    "opentelemetry-exporter-otlp-proto-http>=1.37.0",
    "opentelemetry-sdk>=1.37.0",
    "pygame>=2.6.1",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "rich>=14.1.0",
    "tensorflow>=2.20.0",
]
